backend/requirements.txt
fastapi
uvicorn
pandas
numpy
prophet
scikit-learn
python-multipart
fastapi
uvicorn
pydantic
pandas
tensorflow
numpy
pillow
python-multipart
plotly